requirements.txt
anyio==4.11.0
branca==0.8.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
folium==0.20.0
fonttools==4.60.1
geopandas==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ibm-cos-sdk==2.14.3
ibm-cos-sdk-core==2.14.3
ibm-cos-sdk-s3transfer==2.14.3
ibm_watsonx_ai==1.4.5
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
lomond==0.3.3
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.10.2
numpy==2.3.4
packaging==25.0
pandas==2.2.3
patsy==1.0.2
python-dotenv==1.0.0
pillow==12.0.0
plotly==6.4.0
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6-WebEngine==6.10.0
PyQt6-WebEngine-Qt6==6.10.0
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scipy==1.16.3
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
statsmodels==0.14.5
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xyzservices==2025.10.0
